Move `new_full` to `builder.rs` with zero other changes
